Got jquery pulled into the project, along with several plugins to do table sorting, notifications, and some UI snazziness.
Got a good initial page going, showing a lot of information about the
player's current state.  Next up is to revise the README a bit to plan
out how to proceed.  I slacked a bit on the ajaxy stuff, so the UI is
built by lisp instead of by js, which has me balking on the main plan of
having lisp speak primarily JSON.